introduced setDefaults functions
