+++ internal/app/app_check_clusters.go
@@ -121,12 +121,12 @@ func (p *PolicyAutomationApp) evaluateClusters(regoPackageBases []string) error 
 		return err
 	}
 	clusterData, errors := inputs.GetAllInputsData(p.inputs, clusterIds)
-	if errors != nil && len(errors) > 0 {
+	if len(errors) > 0 {
 		p.out.ErrorPrint("could not fetch the cluster details", errors[0])
 		log.Errorf("could not fetch cluster details: %s", errors[0])
 		return errors[0]
 	}
-	val, err := json.MarshalIndent(clusterData, "", "    ")
+	val, _ := json.MarshalIndent(clusterData, "", "    ")
 	log.Debugf("[DEBUG] cluster: " + string(val))
 
 	evalResults := &evaluationResults{}
